Add presenter. Code review
