+++ src/presenter/index.js
@@ -1,0 +1,5 @@
+import Trip from './trip/trip';
+
+export {
+  Trip
+};
